requirements.txt
django-bootstrap5
django-bootstrap-daterangepicker
Django==3.*
djangorestframework-api-key==2.*
djangorestframework==3.12.2
django-cors-headers==3.6.0
django-compressor
django-heroku
django-sass-processor
libsass
gunicorn
Pillow==8.3.*
psycopg2==2.8.4
python-dotenv==0.12.0
pytz==2019.3
requests==2.24.0
numpy
whitenoise
